python/Basic/Oauth/pyproject.toml
[tool.poetry]
name = "basicoauth"
version = "1.0.0"
description = "Basic OAuth Example"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.6.2"
"fds.sdk.utils" = "^0.9.0"
"fds.sdk.PAEngine" = "^0.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
